Start isolating preprocessor conditionals in functions.  So, we now call str2json_alloc() and json2str_alloc() instead of using macros all along the code.
We also start using the *_alloc() suffix for functions that return a pointer that need to be freed by the caller.  This allowed us to easily spot so pointers that needed to be freed, but were not being so.

FossilOrigin-Name: 4b9489e6632abcd7e20aa23a477846a05edb82ed31985f9baaaa642ded5049ae