Fix scrolling problem.
This bug happened when last message was being shown and the user it tried to scroll up. Instead of scrolling up half of a page (with CTRL-B), janechat showed messages from the first one. This was because top_line is a special case that was not handled correctly.

Besides that, add tests for this case and add a [ctrl] function to virtterm so we can emulate CTRL sequences.

FossilOrigin-Name: d15d0da1b571fc74c9672ca289309d57489bf79cd410325a1a1137ade5b49772